requirements.txt
streamlit
numpy
opencv-python
Pillow
google-generativeai
python-dotenv
streamlit-drawable-canvas
